Add a new function format_string to generate the right class attributes where appropriate for tags with a format (something in parens after the tag), and also handle formats that begin with # by turning them into id instead of class.  Add <a name> tag when an id is specified for a heading so that it can be used for internal links.
Correctly handle an empty or missing \id string.  Add a \packnumber
command analogous to \packbullet.

Add the ability to pass options to pod2thread, and pass the -n option by
default to add a navigation bar to the top of the generated XHTML.

Fix spin -h to use the right path to spin.