Significantly changed the internal processing mechanisms for thread so that the code keeps track of whether each element is block or inline and wraps paragraphs appropriately without special-casing <p> and <pre>.  This involved a significant rewrite of the core processing code and the implementation of each of the command handlers.  This also fixes the fact that some commands output directly and others returned their output, which was causing some whitespace irregularities.
Removed \p; the same thing can now be done with \class.  Removed \packbullet
and \packnumber in favor of \bullet(packed) and \number(packed).  Improved
handling of \quote(broken) and \quote in general for multi-paragraph quotes.

Improved handling of whitespace at the end of the document and with
documents generated by external programs.  Removed the unnecessary $OUT
global variable.  Improved warnings for unknown strings and handling of
unknown commands.  Removed the <?xml-stylesheet> directive since it doesn't
fit with Appendix C XHTML 1.0.