Rewrite how page footers are handled so that the code to generate them can be abstracted out rather than repeated.  Generalize them so that if wanted some of the converted pages can also make use of things like the document Id string (although this isn't the case yet).  Lift the URL to my web tools up to a variable at the top of the script.
Move handling of the spin comment into do_heading where it's used rather
than in the top-level spin routine with a global variable.

Clean up most of the duplicate code between all the routines to call
external converters and standardize the format of a pointer file.  It now
contains the pointer to the source file, any additional options, and the
style sheet to use, rather than taking different formats depending on the
converter run.  Update the documentation and remove the relevant BUGS
entry.

Remove the BUGS entry for \p since that was removed in the last set of
changes.