Fix significant bugs in the previous release that I didn't catch due to making a mistake in testing methods.  This release now correctly reproduces the output of previous releases before I started fiddling around with block and inline scope, but still fixes the bug I was trying to fix.  This part of spin is tricky and complex; I wonder if there's a simpler way of handling it.
Add support for adding <link> tags to the <head> section of every page
duplicating the navigation information in the footer in a more
browser-friendly fashion.  Currently only supports prev, next, up, and top
(first and last could be added but haven't been yet).

If .signature is not found in the current directory, check the top of the
source tree for a .signature file there as well.

Regenerate a page produced by an external program if the pointer file has
changed as well as if the source file has changed.