Rework how block structure elements (<blockquote> and lists) are handled internally so that they can finally be nested.  Maintain a stack of indentation and block structure pairs and maintain it properly on changes of indentation in the source file.
Header recognition is now even more conservative, due to the false
positives from mostly looking at short lines.  In order for a header to be
recognized as such, it now must be outdented, underlined, or in all caps.
(Recognizing the first outdented header of a document is still rather
tricky.)

Only use Id strings that have been expanded, getting rid of the strange
results when converting a document with an Id tag that hasn't been
committed to CVS yet.  Get rid of a few other warnings in unusual
situations.  Simplify the handling of quoted paragraphs.  Allow for From_
lines in documents that begin with mail headers.  Improve the distinction
between mail headers and internal documentation headers.