Run spin-rss on any .rss file found in a directory being processed before processing the contents of that directory.
Support being called with only a source and no output if the source is a
file.  In this case, output is sent to standard output.  In all cases
where the source is a single file, change directories to the location of
that file so that \include will work properly.

Add \strike, \sub, and \sup inline commands.