Add support for Markdown README
Add a new template, tests, and a README.md file for this package.
This is for better documentation display on GitHub.  For the time
being, include the README file in the distribution as well.  (I may
change my mind and drop it later.)

This requires a bit of a rework of section headings, since I don't
want to use all caps with Markdown where we have actual font changes.