Redo the placement of footnotes in text output
When generating text output, put the footnotes containing URLs for
links immediately following the containing paragraph rather than the
end of the text block.  This is both more readable and avoids odd
placement of the footnotes when a template adds further paragraphs to
the end of a text block containing footnotes.