Refactor command handling, add default arguments
Change the App::DocKnot::Generate API to be compatible with the new
dispatch mechanism in App::DocKnot and eliminate the temporary stub
method.  Add a new generate_output() method that writes the output
to a file.

docknot generate now takes the name of the template as its first
positional argument instead of as the -t option.

generate() and therefore the generate subcommand now defaults to
looking for metadata in docs/metadata, and the -m option is now
optional.

The -o option to docknot generate and the output argument to
generate_output() is now optional if there is a default output file
defined for the template.  The two defaults are README for the readme
template and README.md for the readme-md template, both in the current
directory.

Allow all App::DocKnot submodules to croak on error and intercept
those errors and rewrite them into good command errors in App::DocKnot,
making it easier to use consistent error handling between the direct
API and the command-line API.