Separate the testing section for Perl packages
Separate the TESTING section from BUILD AND INSTALLATION for Perl
modules in both README and README.md, and move the Lancaster
instructions to that section.  Add instructions for how to run a
single test with Perl build systems.  Test dependencies for Perl
modules should now move into docs/metadata/test/suffix, the same as
for packages with an Autoconf build system.