Allow the path to Perl to be configured for dist
Allow the path to Perl to be configured in App::DocKnot::Dist.  This
is used primarily for testing so that we can try building a
distribution using Module::Build with the same Perl used to run the
test suite.  This is exposed by the module API but (intentially) not
exposed on the docknot command line.  Thanks, Slaven Rezic.  (#129958)