Remove a distribution directory if it already exists
docknot dist now removes any existing directory when creating a new
distribution.  It overrides permissions if possible while doing this
so that it can clean up from a failed Automake make distcheck, so the
destination directory should be trusted, not one to which an attacker
may be able to write files.