Add support for GitHub Actions workflows
Change the vcs.travis metadata key to vcs.status.travis.  This will
require changes to all templates and package metadata.json files that
use this key.

Add support for generating a build status badge from a GitHub Actions
workflow.  To do this, set vcs.status.workflow to the name of the
workflow that should be referenced for the badge.

Also specify DocKnot's dependencies with a cpanfile.  This makes it
easier to install its dependencies with some tools, such as cpm.