Drop support for support.cpan
Drop support for the support.cpan metadata key, since the CPAN RT
instance is going away.  For packages with support.cpan set, if
vcs.github was set and support.github was not, set support.github to
match vcs.github during a docknot update.

Change DocKnot to point to the GitHub issue tracker and make the
same change to the affected test data.