Overhaul supplemental section handling
Support a test.override metadata key that overrides the Testing
section in README and README.md files entirely, except for the note
about Lancaster Consensus environment variables.

Move readme.sections to just sections, and if it defined a testing
section, move that to test.override.

Add some additional markup to the Markdown version of building
instructions for packages that use Kerberos and Autoconf.