Various documentation cleanups
Fix left-over documentation references to the old metatada path.
Refer to DocKnot rather than App::DocKnot as the canonical name of
the package, and reference the App-DocKnot distribution in the
docknot man page.  Add a few more cross-references to the module
documentation pages.  Flesh out some of the documentation about
tarball signing.  Update the synopsis of the docknot man page.