Add test for spinning a single file
Test spinning a single file with App::DocKnot::Spin and change
just enough to get the test to pass.  Hard-code the old version
number for now, since there doesn't seem to be an immediately
obvious way to get it.  Don't use standard output when spinning
a single file since it confuses prove somehow.