Fix command plumbing for spin and spin-file
Plumb spin and spin-file through App::DocKnot::Command to the
docknot frontend again.  Add documentation for those commands and
their options.  Add a test for the command-line interface to
ensure that plumbing keeps working in the future.

Move the logic to compare a tree of spun files into the
Test::DocKnot::Spin module so that it can be shared between the
command-line test and the API test.

Add the test library modules to POD syntax and spelling checks.