Clean up error reporting, mostly for spin
Standardize on lowercase letters starting error messages.  Enable
autodie for App::DocKnot::Spin (which was already assumed by the
new API, whoops) and clean up a few places where open was used with
the old file handle syntax.

Add a helper function for checked print to a file handle and use it
everywhere.

Only store the input and output path in the object state when
spinning a tree, and adjust the code to not try to call functions
that need it when spinning a single file.

Use the _warning helper method everywhere to include the file and
line information, and add a _fatal helper for the places this was
done with die.  Remove $0 from die invocations, since this is
handled by App::DocKnot::Command.