Fix spin with external converters
The refactorings broke external converters.  Fix that, hopefully,
and add a test for pod2html.  Refactor the division of labor between
the conversion methods and the common method so that it works
without reinvoking spin via $FULLPATH.  Drop the $FULLPATH global
variable.

Update perltidy configuration and apply the reformattings to avoid
some really horrible formatting decisions in some of the new test
code.

Add a dependency on Pod::Thread, which is currently only used for
the external script but will eventually be used for its API.