Fix lifting of \class in paragraphs
Thread supports a feature where \class surrounding a paragraph is
lifted into the paragraph tag, but it was firing when \class
surrounded the start and end of the paragraph but not all of its
text.  Restore the check that ensures it surrounds the entire
text (which is not 100% correct because we're not doing full HTML
parsing, but should be good enough).