Add *.spin pointers and Markdown support
Add a new *.spin input file for docknot spin that points to an external
file with instructions for how to convert it to HTML.  Via this
mechanism, support formatting Markdown files as HTML via docknot spin.
This support requires the pandoc program be installed.  The path to
pandoc may be specified in the DocKnot global configuration file.

Remove an extra blank line after the leading navbar in output
converted from thread.