Fix docknot spin requirement for package metadata
Now that docknot spin loads App::DocKnot::Config in order to get
the global config, too-aggressive error checking in the
constructor caused it to fail unless run from a directory
containing a package metadata file.  Fix this by not checking that
the metadata path is valid in the constructor.  YAML::XS will
produce a reasonable error when trying to open a nonexistent file.