Rename orphaned to unmaintained
In the DocKnot metadata, rename orphaned to unmaintained.  Don't
try to update old YAML, since there isn't good support for that
now, but fix the conversion from JSON.  Change the output language
in the templates accordingly.

Orphaned is the historic term in free software, but it has a lot
of emotional overtones that aren't relevant to software.
Unmaintained is more precise and relies less on metaphor.