Convert faq2html to a module
Merge faq2html into this package as App::DocKnot::Spin::Text, and
add a new command, docknot spin-text, which takes roughly the same
options as faq2html.  This adds the ability to convert text files in
my personal format into HTML.

Add a basic test suite, which currently does not adjust for dates
correctly.  The documentation of the module is also entirely wrong,
and some tests are disabled until the module source can be refactored.